지면 추가, 플레이어 점프 구현 + 이동생성자로 컴포넌트 추가하면 문제가 많아서 삭제

+++ IJumpable.h
@@ -5,9 +5,9 @@
 class IJumpable
 {
 protected:
-	int jumpPower;
+	float jumpPower;
 
-	IJumpable(int power = 700) : jumpPower(power) {}
+	IJumpable(float power = 700.f) : jumpPower(power) {}
 
 public:
 	virtual void Jump() = 0;
